Many, many bugfixes.

+++ brg/physics/lensing/magnification/Schechter_like_functor.cpp
@@ -30,18 +30,18 @@
 
 namespace brgastro {
 
-double Schechter_like_functor::operator()( const double & in_param,
+long double Schechter_like_functor::operator()( const long double & in_param,
 			const bool silent) const
 {
 	const double mag_jump_limit = 23.;
 
-	const double x = std::pow(10,0.4*(m_star()-in_param));
-	double result = 0.4*std::log(10.)*N_scale()*std::pow(x,alpha()+1)*
+	const long double x = std::pow(10,0.4*(m_star()-in_param));
+	long double result = 0.4*std::log(10.)*N_scale()*std::pow(x,alpha()+1)*
 			std::exp(-std::pow(x,mag_lower_lim_sharpness()));
 
 	if(in_param>=mag_jump_limit) result += mag23_jump();
 
-	const double xh = std::pow(10,0.4*(in_param-mag_upper_lim()));
+	const long double xh = std::pow(10,0.4*(in_param-mag_upper_lim()));
 
 	result *= std::exp(-std::pow(xh,mag_upper_lim_sharpness()));
